+++ test/client/self_issued.test.js
@@ -1,7 +1,7 @@
 const { expect } = require('chai');
 const nock = require('nock');
 const timekeeper = require('timekeeper');
-const jose2 = require('jose2');
+const jose = require('jose');
 
 const { Issuer } = require('../../lib');
 
@@ -35,29 +35,33 @@ describe('Validating Self-Issued OP responses', () => {
     Object.assign(this, { issuer, client });
   });
 
-  const idToken = (claims = {}) => {
-    const jwk = jose2.JWK.generateSync('EC');
-    return jose2.JWT.sign(
-      {
-        sub_jwk: jwk.toJWK(),
-        sub: jwk.thumbprint,
-        ...claims,
-      },
-      jwk,
-      { expiresIn: '2h', issuer: 'https://self-issued.me', audience: 'https://rp.example.com/cb' },
-    );
-  };
+  async function idToken(claims = {}) {
+    const kp = await jose.generateKeyPair('ES256', { extractable: true });
+    const jwk = await jose.exportJWK(kp.publicKey);
+    const sub = await jose.calculateJwkThumbprint(jwk);
+    return await new jose.SignJWT({
+      sub_jwk: jwk,
+      sub,
+      ...claims,
+    })
+      .setIssuedAt()
+      .setProtectedHeader({ alg: 'ES256' })
+      .setIssuer('https://self-issued.me')
+      .setAudience('https://rp.example.com/cb')
+      .setExpirationTime('2h')
+      .sign(kp.privateKey);
+  }
 
   describe('consuming an ID Token response', () => {
-    it('consumes a self-issued response', function () {
+    it('consumes a self-issued response', async function () {
       const { client } = this;
-      return client.callback(undefined, { id_token: idToken() });
+      return client.callback(undefined, { id_token: await idToken() });
     });
 
-    it('expects sub_jwk to be in the ID Token claims', function () {
+    it('expects sub_jwk to be in the ID Token claims', async function () {
       const { client } = this;
       return client
-        .callback(undefined, { id_token: idToken({ sub_jwk: undefined }) })
+        .callback(undefined, { id_token: await idToken({ sub_jwk: undefined }) })
         .then(fail, (err) => {
           expect(err.name).to.equal('RPError');
           expect(err.message).to.equal('missing required JWT property sub_jwk');
@@ -65,10 +69,10 @@ describe('Validating Self-Issued OP responses', () => {
         });
     });
 
-    it('expects sub_jwk to be a public JWK', function () {
+    it('expects sub_jwk to be a public JWK', async function () {
       const { client } = this;
       return client
-        .callback(undefined, { id_token: idToken({ sub_jwk: 'foobar' }) })
+        .callback(undefined, { id_token: await idToken({ sub_jwk: 'foobar' }) })
         .then(fail, (err) => {
           expect(err.name).to.equal('RPError');
           expect(err.message).to.equal('failed to use sub_jwk claim as an asymmetric JSON Web Key');
@@ -76,13 +80,15 @@ describe('Validating Self-Issued OP responses', () => {
         });
     });
 
-    it('expects sub to be the thumbprint of the sub_jwk', function () {
+    it('expects sub to be the thumbprint of the sub_jwk', async function () {
       const { client } = this;
-      return client.callback(undefined, { id_token: idToken({ sub: 'foo' }) }).then(fail, (err) => {
-        expect(err.name).to.equal('RPError');
-        expect(err.message).to.equal('failed to match the subject with sub_jwk');
-        expect(err).to.have.property('jwt');
-      });
+      return client
+        .callback(undefined, { id_token: await idToken({ sub: 'foo' }) })
+        .then(fail, (err) => {
+          expect(err.name).to.equal('RPError');
+          expect(err.message).to.equal('failed to match the subject with sub_jwk');
+          expect(err).to.have.property('jwt');
+        });
     });
   });
 });